refactor: extract wave2 selection logic from investigate/workflow.ts
